style(esl-event-listener): refactor types and remove incorrect tests
